requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.6.4
propcache==0.3.2
python-dateutil==2.9.0.post0
python-telegram-bot==22.3
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
